Updated the images api to the new app route
